Move logic out of scripts into Makefile (#122)
The motivation behind this is to keep the logic within the scripts more straight-forward so that they can be run directly by the users. Magic to extract parameters and so forth is more appropriate to be in the Makefile.

The key observation here is that the functionality is already available from within the deployment container. Thus we can re-use to deployer container for parameter manipulation. This fixes #106. We are able to thin out a few scripts.

Also clean up python tests.